# Conflicts:
#	The_Reviewer_Project/src/test/java/academy/mindswap/mockdata/MockedData.java

+++ The_Reviewer_Project/src/main/java/academy/mindswap/commands/ReviewDto.java
@@ -20,5 +20,5 @@ public class ReviewDto {
     @NotNull
     @Max(5)
     @Min(1)
-    private Integer localRating;
+    private Float localRating;
 }
